+++ src/main/java/game/field/Ladder.java
@@ -1,4 +1,6 @@
-package ladder;
+package game.field;
+
+import game.fieldComponent.Row;
 
 public class Ladder {
