+++ src/main/java/game/pipe/util/CustomException.java
@@ -1,4 +1,4 @@
-package ladder.util;
+package game.pipe.util;
 
 public class CustomException extends RuntimeException {
     private final ErrorCode errorCode;
