+++ src/main/java/game/pipe/util/ErrorCode.java
@@ -1,4 +1,4 @@
-package ladder.util;
+package game.pipe.util;
 
 public enum ErrorCode {
     LACK_PARTICIPANTS("게임의 참여자 수는 1명 이상이어야 합니다."),
