+++ src/main/java/game/pipe/util/PipeConstant.java
@@ -1,4 +1,4 @@
-package ladder.util;
+package game.pipe.util;
 
 public final class PipeConstant {
